# docs/swagger.yaml
definitions:
  content.CreateContentRequest:
    properties:
      contentType:
        type: string
      desc:
        type: string
      genre:
        type: string
      releaseDate:
        type: string
      title:
        type: string
    type: object
  content.CreateContentResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.ContentResponse'
      msg:
        type: string
    type: object
  content.DeleteContentResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.ContentResponse'
      msg:
        type: string
    type: object
  content.ErrorResponse:
    properties:
      msg:
        type: string
    type: object
  content.ListContentsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.ContentResponse'
        type: array
      msg:
        type: string
    type: object
  content.ShowContentResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.ContentResponse'
      msg:
        type: string
    type: object
  content.UpdateContentsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.ContentResponse'
        type: array
      msg:
        type: string
    type: object
  content.updateContentRequest:
    properties:
      contentType:
        type: string
      desc:
        type: string
      genre:
        type: string
      releaseDate:
        type: string
      title:
        type: string
    type: object
  schemas.ContentResponse:
    properties:
      contentType:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      desc:
        type: string
      genre:
        type: string
      id:
        type: integer
      realeaseDate:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /Content:
    delete:
      consumes:
      - application/json
      description: Delete a new Media Content
      parameters:
      - description: Content Identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.DeleteContentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/content.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/content.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/content.ErrorResponse'
      summary: Delete Content
      tags:
      - Media Contents
    get:
      consumes:
      - application/json
      description: Show a Media Content
      parameters:
      - description: Content Identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.ShowContentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/content.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/content.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/content.ErrorResponse'
      summary: Show Content
      tags:
      - Media Contents
    post:
      consumes:
      - application/json
      description: Create a new Media Content
      parameters:
      - description: Content data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/content.CreateContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.CreateContentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/content.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/content.ErrorResponse'
      summary: Create Content
      tags:
      - Media Contents
    put:
      consumes:
      - application/json
      description: Update a Media Content
      parameters:
      - description: Content Identification
        in: query
        name: id
        required: true
        type: string
      - description: Content data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/content.updateContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.UpdateContentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/content.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/content.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/content.ErrorResponse'
      summary: Update Content
      tags:
      - Media Contents
  /Contents:
    get:
      consumes:
      - application/json
      description: list all Media Contents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/content.ListContentsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/content.ErrorResponse'
      summary: List Contents
      tags:
      - Media Contents
swagger: "2.0"